refactor: controller와 repository 디렉터리 구분
