+++ src/main/java/roomescape/repository/AdminRepository.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.repository;
 
 import java.util.HashMap;
 import java.util.Map;
